+++ components/atom/input/src/Input/index.js
@@ -1,7 +1,7 @@
 import AddonHoc from './Features/Addon'
 import IconHoc from './Features/Icon'
 import WithButtonHoc from './Features/WithButton'
-import Component, {inputSizes} from './Component'
+import Component, {inputStates, inputSizes} from './Component'
 
 export default WithButtonHoc(AddonHoc(IconHoc(Component)))
-export {inputSizes}
+export {inputSizes, inputStates}