+++ components/molecule/autosuggest/src/index.js
@@ -15,9 +15,10 @@ const BASE_CLASS = `sui-MoleculeAutosuggest`
 const CLASS_FOCUS = `${BASE_CLASS}--focus`
 const CLASS_DISABLED = `${BASE_CLASS}--disabled`
 
-const ERROR_STATES = {
+const AUTOSUGGEST_STATES = {
   ERROR: 'error',
-  SUCCESS: 'success'
+  SUCCESS: 'success',
+  ALERT: 'alert'
 }
 
 const getIsTypeableKey = key => {
@@ -43,7 +44,8 @@ const MoleculeAutosuggest = ({multiselection, ...props}) => {
     keysCloseList,
     keysSelection,
     disabled,
-    errorState
+    errorState,
+    state
   } = props
 
   const refMoleculeAutosuggest = useRef(
@@ -66,8 +68,9 @@ const MoleculeAutosuggest = ({multiselection, ...props}) => {
 
   const className = cx(
     BASE_CLASS,
-    errorState && `${BASE_CLASS}--${ERROR_STATES.ERROR}`,
-    errorState === false && `${BASE_CLASS}--${ERROR_STATES.SUCCESS}`,
+    errorState && `${BASE_CLASS}--${AUTOSUGGEST_STATES.ERROR}`,
+    errorState === false && `${BASE_CLASS}--${AUTOSUGGEST_STATES.SUCCESS}`,
+    state && `${BASE_CLASS}--${state}`,
     {
       [CLASS_FOCUS]: focus,
       [CLASS_DISABLED]: disabled
@@ -78,7 +81,7 @@ const MoleculeAutosuggest = ({multiselection, ...props}) => {
     const {current: domMoleculeAutosuggest} = refMoleculeAutosuggest
     onToggle(ev, {isOpen: false})
     if (multiselection) onChange(ev, {value: ''})
-    domMoleculeAutosuggest.focus()
+    domMoleculeAutosuggest && domMoleculeAutosuggest.focus()
     setFocus(false)
     ev.preventDefault()
     ev.stopPropagation()
@@ -252,7 +255,13 @@ MoleculeAutosuggest.propTypes = {
   tabIndex: PropTypes.number,
 
   /** true = error, false = success, null = neutral */
-  errorState: PropTypes.bool
+  errorState: PropTypes.bool,
+
+  /* Will set a red/green/orange border if set to 'error' / 'success' / 'alert' */
+  state: PropTypes.oneOf(Object.values(AUTOSUGGEST_STATES)),
+
+  /* Button prop to be passe down to the input field */
+  rightButton: PropTypes.node
 }
 
 MoleculeAutosuggest.defaultProps = {
@@ -267,3 +276,4 @@ MoleculeAutosuggest.defaultProps = {
 
 export default withOpenToggle(MoleculeAutosuggest)
 export {SIZES as MoleculeAutosuggestDropdownListSizes}
+export {AUTOSUGGEST_STATES as MoleculeAutosuggestStates}