+++ components/molecule/field/src/index.js
@@ -19,6 +19,7 @@ const MoleculeField = ({
   reverse,
   errorText,
   successText,
+  alertText,
   label,
   useContrastLabel,
   helpText,
@@ -45,6 +46,7 @@ const MoleculeField = ({
   if (useContrastLabel) {
     typeValidationLabel = AtomLabelTypes.CONTRAST
   }
+
   if (errorText) {
     statusValidationText = errorText
     typeValidationLabel = AtomLabelTypes.ERROR
@@ -57,20 +59,28 @@ const MoleculeField = ({
     typeValidationText = AtomValidationTextTypes.SUCCESS
   }
 
+  if (alertText) {
+    statusValidationText = alertText
+    typeValidationLabel = AtomLabelTypes.ALERT
+    typeValidationText = AtomValidationTextTypes.ALERT
+  }
+
   return (
     <div className={className}>
-      <div className={CLASS_LABEL_CONTAINER}>
-        {inline && extendedChildren}
-        <AtomLabel
-          type={typeValidationLabel}
-          name={name}
-          text={label}
-          onClick={onClickLabel}
-        />
-      </div>
+      {label && (
+        <div className={CLASS_LABEL_CONTAINER}>
+          {inline && extendedChildren}
+          <AtomLabel
+            type={typeValidationLabel}
+            name={name}
+            text={label}
+            onClick={onClickLabel}
+          />
+        </div>
+      )}
       <div className={CLASS_INPUT_CONTAINER}>
         {!inline && extendedChildren}
-        {(successText || errorText) && (
+        {typeValidationText && (
           <AtomValidationText
             type={typeValidationText}
             text={statusValidationText}
@@ -89,7 +99,7 @@ MoleculeField.propTypes = {
   children: PropTypes.any,
 
   /** Text to be displayed as label of the textarea */
-  label: PropTypes.string.isRequired,
+  label: PropTypes.string,
 
   /** If true it will set the label type to 'CONTRAST' */
   useContrastLabel: PropTypes.bool,
@@ -105,6 +115,9 @@ MoleculeField.propTypes = {
 
   /** Error message to display when error state  */
   errorText: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
+
+  /** Error message to display when alert state  */
+  alertText: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
 
   /** Help Text to display */
   helpText: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),