+++ components/molecule/select/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.24.0 (2020-02-03)
+
+
+### Features
+
+* add state prop ([1834386](https://github.com/SUI-Components/sui-components/commit/18343869dfc67288a1526271f32d76e173d68a9b))
+
+
+
+# 1.23.0 (2020-01-28)
+
+
+### Bug Fixes
+
+* moleculeSelectSingleSelection refMoleculeSelect.current.focus() crash fix ([87e8f4e](https://github.com/SUI-Components/sui-components/commit/87e8f4e77a4c0edc5059b672319db247d549c5ed))
+
+
+
 # 1.22.0 (2020-01-21)